+++ s/tools/fancy_event_listener.ts
@@ -1,5 +1,0 @@
-
-export interface FancyEventListener<E extends Event> extends AddEventListenerOptions {
-	handleEvent(e: E): void
-}
-