+++ s/tools/lit_listener.ts
@@ -1,0 +1,9 @@
+
+export interface LitListener<E extends Event = Event> extends AddEventListenerOptions {
+	handleEvent(e: E): void
+}
+
+export function litListener<E extends Event = Event>(o: LitListener<E>) {
+	return o
+}
+